Extract class with Scheduling Sample Configurations.
